refractor: split gitea_test.go into focused files
rh-pre-commit.version: 2.4.0
rh-pre-commit.check-secrets: ENABLED
